WEB-1112: Break product item out into partial
